indexado de nodos individualemente, si existe un indexado anterior, no se elimina en este caso

+++ series_tiempo_ar_api/apps/metadata/indexer/metadata_indexer.py
@@ -20,24 +20,28 @@ class MetadataIndexer:
         self.elastic: Elasticsearch = connections.get_connection()
         self.task = task
 
-    def update_alias(self, index_name):
+    def update_alias(self, index_name, new_index=True):
         if not self.elastic.indices.exists_alias(name=constants.METADATA_ALIAS):
             self.elastic.indices.put_alias(index_name, constants.METADATA_ALIAS)
             return
-        indices = self.elastic.indices.get_alias(name=constants.METADATA_ALIAS).keys()
 
         actions = [
             {"add": {"index": index_name, "alias": constants.METADATA_ALIAS}},
         ]
 
-        for old_index in indices:
-            actions.append({"remove_index": {"index": old_index}})
+        if new_index:
+            indices = self.elastic.indices.get_alias(name=constants.METADATA_ALIAS).keys()
+            for old_index in indices:
+                actions.append({"remove_index": {"index": old_index}})
 
         self.elastic.indices.update_aliases({
             "actions": actions
         })
 
-    def run(self, nodes, new_index=True):
+    def run(self, nodes, single_node=False):
+        new_index = \
+            not single_node or \
+            not self.elastic.indices.exists_alias(name=constants.METADATA_ALIAS)
         index = get_random_index_name() if new_index else constants.METADATA_ALIAS
         index_created = False
         for node in nodes:
@@ -52,22 +56,19 @@ class MetadataIndexer:
                 IndexMetadataTask.info(self.task,
                                        u'Error en la lectura del catálogo {}: {}'.format(node.catalog_id, e))
 
-        if not new_index:
-            return
-
         if index_created:
             self.elastic.indices.forcemerge(index=index)
-            self.update_alias(index)
-        else:
-            if self.elastic.indices.exists(index):
-                self.elastic.indices.delete(index)
+            self.update_alias(index, new_index=new_index)
+        elif self.elastic.indices.exists(index):
+            self.elastic.indices.delete(index)
 
 
 # pylint: disable=W0613
 @job('meta_indexing', timeout=10000)
 def run_metadata_indexer(task):
-    nodes = [task.node] if task.node else Node.objects.filter(indexable=True)
-    MetadataIndexer(task).run(nodes, new_index=not bool(task.node))
+    single_node = bool(task.node)
+    nodes = [task.node] if single_node else Node.objects.filter(indexable=True)
+    MetadataIndexer(task).run(nodes, single_node=single_node)
     update_units()
     task.refresh_from_db()
     task.status = task.FINISHED
